Sell Item
Uhh, kind of embarrassed that I couldn't get this done this evening. So I've tried to get the date picker working but it doesn't and I've realised before I can make it work I need to know who is using the form- i.e. who is logged in and I think that means you need to navigate to the form and remain logged in