update watchlist messages and view item
I've put a note on line 59 about some changes to view item I made because it wasn't working on wamp. However it doesn't appear to be a problem on amazon, so you don't have to use it. The main change here was that start price was shown if no bids

watchlist messages now seem to send 1 message per bidder and watchlister.